Modulo_9/ponderada4_metabase/requirements.txt
paho-mqtt==2.0.0
pytest==8.0.1
